pass placeholder to worker
