+++ scripts/headerfooter-worker.js
@@ -3,16 +3,17 @@
  * @param {string} content string of html with relative urls
  * @returns the string with absolute urls
  */
-function resolveRelativeURLs(content) {
+function resolveRelativeURLs(content, privacyIcon, localPrivacyIcon) {
   const baseUrl = 'https://walgreens.com';
 
   // Use a regular expression to find relative links (starting with "/")
   const relativeLinkRegex = /(?:href|action)="(?!\/images\/)(\/[^"]+)"/g;
-  const absoluteContent = content.replace(relativeLinkRegex, (match, relativePath) => {
+  let absoluteContent = content.replace(relativeLinkRegex, (match, relativePath) => {
     // Combine the base URL and the relative path to create an absolute URL
     const absoluteUrl = `${baseUrl}${relativePath}`;
     return `href="${absoluteUrl}"`;
   });
+  absoluteContent = absoluteContent.replace(privacyIcon, localPrivacyIcon);
   return absoluteContent;
 }
 
@@ -21,7 +22,7 @@ onmessage = async (e) => {
     header: 'https://www.walgreens.com/common/v1/headerui',
     footer: 'https://www.walgreens.com/common/v1/footerui',
   };
-  const { source } = e.data;
+  const { source, privacyIcon, localPrivacyIcon } = e.data;
   const resp = await fetch(APIs[source]);
 
   if (!resp.ok) {
@@ -30,7 +31,7 @@ onmessage = async (e) => {
   }
 
   const jsonData = await resp.json();
-  jsonData.content = resolveRelativeURLs(jsonData.content);
+  jsonData.content = resolveRelativeURLs(jsonData.content, privacyIcon, localPrivacyIcon);
 
   postMessage({
     source,
